+++ components/Templates/Map/RightPanel/PlaceDetail.tsx
@@ -57,10 +57,10 @@ const PlaceDetail: React.FC<{
                   key={post.id}
                   className="flex flex-col w-full h-64 text-left bg-white border border-gray-200 rounded-lg hover:border-gray-400 hover:shadow-sm"
                 >
-                  {post.url && (
+                  {post.photos && (
                     <div className="relative w-full h-full rounded-t-lg cursor-pointer bg-secondary">
                       <Image
-                        src={UrlPrefix + post.url[0]}
+                        src={UrlPrefix + post.photos[0].hash}
                         layout="fill"
                         objectFit="cover"
                         objectPosition="center"
