+++ lib/graphql/auth.graphql
@@ -1,0 +1,23 @@
+mutation verifyToken($token: String!) {
+  verifyToken(token: $token)
+}
+
+mutation login($input: LoginInput!) {
+  login(input: $input)
+}
+
+mutation googleAuth($googleJWT: String!) {
+  googleAuth(googleJWT: $googleJWT)
+}
+
+mutation forgotPassword($login: String!) {
+  forgotPassword(login: $login)
+}
+
+mutation resetPassword($input: ResetPasswordInput!) {
+  resetPassword(input: $input)
+}
+
+mutation createUser($input: CreateUserInput!) {
+  createUser(input: $input)
+}
