+++ lib/graphql/user.graphql
@@ -1,15 +1,3 @@
-mutation verifyToken($token: String!) {
-  verifyToken(token: $token)
-}
-
-mutation login($input: LoginInput!) {
-  login(input: $input)
-}
-
-mutation googleAuth($googleJWT: String!) {
-  googleAuth(googleJWT: $googleJWT)
-}
-
 query suggestedUsers {
   suggestedUsers {
     id
@@ -17,14 +5,6 @@ query suggestedUsers {
     lastName
     username
   }
-}
-
-mutation forgotPassword($login: String!) {
-  forgotPassword(login: $login)
-}
-
-mutation resetPassword($input: ResetPasswordInput!) {
-  resetPassword(input: $input)
 }
 
 query user($username: String, $id: Int) {
@@ -95,8 +75,4 @@ query Me {
       newsletter
     }
   }
-}
-
-mutation createUser($input: CreateUserInput!) {
-  createUser(input: $input)
 }
