+++ pages/forgot-password.tsx
@@ -1,7 +1,7 @@
 import { Input } from '@components/Elements';
 import Button from '@components/Elements/Buttons/Button';
 import AuthLayout from '@components/Layouts/Auth';
-import { useForgotPasswordMutation } from '@graphql/user.graphql';
+import { useForgotPasswordMutation } from '@graphql/auth.graphql';
 import { RedirectLogged } from '@lib/functions/ServerSideProps';
 import ForgotPasswordFormInputs from '@lib/types/forms/forgotPasswordFormInputs';
 import Link from 'next/link';
@@ -10,16 +10,15 @@ import { SubmitHandler, useForm } from 'react-hook-form';
 import { toast } from 'react-hot-toast';
 import { useTranslation } from 'react-i18next';
 
-
 const ForgotPassword: React.FC = () => {
   const [forgotPasswordMutation] = useForgotPasswordMutation(); // forgot password mutation (creates token and sends email)
-  const [loading, setLoading] = useState(false);  // loading after submit
-  const { t } = useTranslation();     // translation
+  const [loading, setLoading] = useState(false); // loading after submit
+  const { t } = useTranslation(); // translation
 
   const {
     register,
     handleSubmit,
-    formState: { },
+    formState: {},
   } = useForm<ForgotPasswordFormInputs>();
 
   const onSubmit: SubmitHandler<ForgotPasswordFormInputs> = async (data) => {
@@ -37,7 +36,6 @@ const ForgotPassword: React.FC = () => {
     setLoading(false);
   };
 
-
   return (
     <AuthLayout
       head={{
@@ -53,7 +51,6 @@ const ForgotPassword: React.FC = () => {
         },
       }}
     >
-
       <form onSubmit={handleSubmit(onSubmit)} className="flex flex-col gap-2">
         <Input
           label={t('username or email')}
@@ -78,6 +75,5 @@ const ForgotPassword: React.FC = () => {
 };
 
 export const getServerSideProps = RedirectLogged;
-
 
 export default ForgotPassword;
